sam4l: update to moved deferred_call
