+++ chips/sam4l/src/chip.rs
@@ -6,11 +6,12 @@ use ast;
 use cortexm4;
 use crccu;
 use dac;
+use deferred_call_tasks::Task;
 use dma;
 use flashcalw;
 use gpio;
-use helpers::{DeferredCall, Task};
 use i2c;
+use kernel::common::deferred_call;
 use kernel::support;
 use kernel::Chip;
 use pm;
@@ -74,7 +75,7 @@ impl Chip for Sam4l {
 
         unsafe {
             loop {
-                if let Some(task) = DeferredCall::next_pending() {
+                if let Some(task) = deferred_call::DeferredCall::next_pending() {
                     match task {
                         Task::Flashcalw => flashcalw::FLASH_CONTROLLER.handle_interrupt(),
                     }
@@ -151,7 +152,7 @@ impl Chip for Sam4l {
     }
 
     fn has_pending_interrupts(&self) -> bool {
-        unsafe { cortexm4::nvic::has_pending() || DeferredCall::has_tasks() }
+        unsafe { cortexm4::nvic::has_pending() || deferred_call::has_tasks() }
     }
 
     fn mpu(&self) -> &cortexm4::mpu::MPU {
